update refundAmount and price to use double

+++ src/main/kotlin/com/doni/kotlinrestreturnman/model/GetReturnOrderResponse.kt
@@ -1,7 +1,7 @@
 package com.doni.kotlinrestreturnman.model
 
 class GetReturnOrderResponse (
-            val refundAmount: Int,
+            val refundAmount: Double,
             val status: ReturnOrderStatus,
             val items: List<GetItemsResponse>
 )
